Simplified contracts-info overriding
