+++ voting/orbs-pos-data/src/declarations-index.ts
@@ -2,4 +2,4 @@ export * from './index';
 export { IGuardianInfo } from './interfaces/IGuardianInfo';
 export { IOrbsPOSDataService } from './interfaces/IOrbsPOSDataService';
 export { IStakingService } from './interfaces/IStakingService';
-export { IOrbsPosContractsAddresses } from './services/EthereumClientService';
+export { IOrbsPosContractsAddresses } from './contracts-adresses';
